dom/media/webaudio/test/test_audioBufferSourceNodeEnded.html

<!DOCTYPE HTML>
<html>
<head>
  <title>Test ended event on AudioBufferSourceNode</title>
  <script type="text/javascript" src="/tests/SimpleTest/SimpleTest.js"></script>
  <script type="text/javascript" src="webaudio.js"></script>
  <link rel="stylesheet" type="text/css" href="/tests/SimpleTest/test.css" />
</head>
<body>
<pre id="test">
<script class="testbody" type="text/javascript">

SimpleTest.waitForExplicitFinish();
addLoadEvent(function() {
  var context = new AudioContext();
  var buffer = context.createBuffer(1, 2048, context.sampleRate);
  for (var i = 0; i < 2048; ++i) {
    buffer.getChannelData(0)[i] = Math.sin(440 * 2 * Math.PI * i / context.sampleRate);
  }

  var source = context.createBufferSource();

  source.onended = function(e) {
    is(e.target, source, "Correct target for the ended event");
    SimpleTest.finish();
  };

  source.start(0);
  source.buffer = buffer;
  source.connect(context.destination);
});

</script>
</pre>
</body>
</html>
